linkedin/requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.10
langchain-openai>=0.0.5
python-dotenv>=1.0.0
openai>=1.0.0
typing-extensions>=4.5.0 